requirements.txt
# Hassett Forecasting System Requirements

# Core Data Science
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Time Series Forecasting
statsmodels>=0.14.0
prophet>=1.1.5

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Database
# sqlite3 is built-in with Python, no need to install
databricks-sql-connector>=4.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
joblib>=1.3.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.2.0
